requirements.txt
Click==7.0
python-twitter==3.5
stravaio==0.0.9
